Rename AdjacencyStore to MutableSegment
