+++ src/test/java/de/hpi/msd/salsa/store/index/MutableSegmentTest.java
@@ -9,23 +9,23 @@ import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
 
-class WriteSegmentTest {
-    private WriteSegment writeSegment;
+class MutableSegmentTest {
+    private MutableSegment mutableSegment;
 
     @BeforeEach
     void setUp() {
-        writeSegment = new WriteSegment(1000);
+        mutableSegment = new MutableSegment(1000);
     }
 
     @Test
     void shouldInsertNodes() {
         long nodeId = 2;
 
-        writeSegment.addEdge(nodeId, 4, 2);
-        writeSegment.addEdge(nodeId, 5, 2);
-        writeSegment.addEdge(nodeId, 6, 2);
+        mutableSegment.addEdge(nodeId, 4, 2);
+        mutableSegment.addEdge(nodeId, 5, 2);
+        mutableSegment.addEdge(nodeId, 6, 2);
 
-        List<Long> targetNodes = writeSegment.getTargetNodes(nodeId);
+        List<Long> targetNodes = mutableSegment.getTargetNodes(nodeId);
         Assertions.assertEquals(targetNodes, Arrays.asList(4L, 5L, 6L));
     }
 
@@ -36,11 +36,11 @@ class WriteSegmentTest {
         List<Long> expectedNodes = new ArrayList<>(nodes);
 
         for (long targetNodeId = 0; targetNodeId < nodes; targetNodeId++) {
-            writeSegment.addEdge(nodeId, targetNodeId, 1);
+            mutableSegment.addEdge(nodeId, targetNodeId, 1);
             expectedNodes.add(targetNodeId);
         }
 
-        List<Long> targetNodes = writeSegment.getTargetNodes(nodeId);
+        List<Long> targetNodes = mutableSegment.getTargetNodes(nodeId);
         Assertions.assertEquals(expectedNodes, targetNodes);
     }
 
@@ -54,12 +54,12 @@ class WriteSegmentTest {
 
         for (long sourceNodeId = 0; sourceNodeId < 10; sourceNodeId++) {
             for (long targetNodeId = 0; targetNodeId < 100; targetNodeId++) {
-                writeSegment.addEdge(sourceNodeId, targetNodeId, 1);
+                mutableSegment.addEdge(sourceNodeId, targetNodeId, 1);
             }
         }
 
         for (long sourceNodeId = 0; sourceNodeId < 10; sourceNodeId++) {
-            final List<Long> targetNodes = writeSegment.getTargetNodes(sourceNodeId);
+            final List<Long> targetNodes = mutableSegment.getTargetNodes(sourceNodeId);
             Assertions.assertEquals(expectedNodes, targetNodes);
         }
     }
